make introspection method package private, move tests
